Minor refactor to allow running with 'python -m ...'
